chore: move KeyEvents shim to core and use in calendar/overlays
