+++ packages/core/package.json
@@ -26,7 +26,7 @@
     "docs",
     "src",
     "stories",
-    "test",
+    "test-helpers",
     "translations",
     "*.js"
   ],
