+++ packages/overlays/src/BaseOverlayController.js
@@ -1,4 +1,5 @@
 import { render, html } from '@lion/core';
+import '@lion/core/src/differentKeyEventNamesShimIE.js';
 import { containFocus } from './utils/contain-focus.js';
 
 /**
